<fix>[ha]: support iscsi device mapper
support attach guest tools

Resolves: SUG-622
